moving labels specs to projects_spec file
