+++ spec/requests/api/projects_spec.rb
@@ -13,6 +13,7 @@ describe API::API do
   let(:snippet) { create(:project_snippet, author: user, project: project, title: 'example') }
   let(:users_project) { create(:users_project, user: user, project: project, project_access: UsersProject::MASTER) }
   let(:users_project2) { create(:users_project, user: user3, project: project, project_access: UsersProject::DEVELOPER) }
+  let(:issue_with_labels) { create(:issue, author: user, assignee: user, project: project, :label_list => "label1, label2") }
 
   describe "GET /projects" do
     before { project }
@@ -630,6 +631,18 @@ describe API::API do
         delete api("/projects/1328", admin)
         response.status.should == 404
       end
+    end
+  end
+
+   describe "GET /projects/:id/labels" do
+    before { issue_with_labels }
+
+    it "should return project labels" do
+      get api("/projects/#{project.id}/labels", user)
+      response.status.should == 200
+      json_response.should be_an Array
+      json_response.first['name'].should == issue_with_labels.labels.first.name
+      json_response.last['name'].should == issue_with_labels.labels.last.name
     end
   end
 end
